user import fixed and moved storage to io
