+++ datasetinsights/datasets/nyudepth.py
@@ -15,7 +15,7 @@ import pandas as pd
 from PIL import Image
 
 import datasetinsights.constants as const
-from datasetinsights.storage.gcs import GCSClient
+from datasetinsights.io.gcs import GCSClient
 
 from .base import Dataset
