+++ datasetinsights/stats/__init__.py
@@ -1,0 +1,9 @@
+from .visualization.plots import (
+    bar_plot,
+    grid_plot,
+    histogram_plot,
+    plot_bboxes,
+    rotation_plot,
+)
+
+__all__ = [bar_plot, grid_plot, histogram_plot, plot_bboxes, rotation_plot]
